pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sudolabs"
version = "1.0.0"
description = "SudoLabs - Terminal Cybersecurity Hacking Lab"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SudoLabs Team"}
]
keywords = ["cybersecurity", "hacking", "lab", "ctf", "pentesting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Education",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "docker>=7.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "anthropic>=0.30.0",
]

[project.scripts]
sudolabs = "sudolabs.cli:app"

[tool.setuptools.packages.find]
include = ["sudolabs*"]
